refactor(TeacherProjectService): Move constraint-related functions to ConstraintService
